feat(controller): 添加多个新功能的 Servlet 实现
- 新增数据统计、系统日志、删除操作、详情查看等多方面功能- 实现了学生、课程、班级等信息的详细查看和修改
- 添加了成绩更新和请假审批相关功能
- 优化了错误处理和用户提示信息

+++ src/main/java/com/azyasaxi/model/CreditSummary.java
@@ -1,48 +1,51 @@
 package com.azyasaxi.model;
 
+import java.math.BigDecimal; // 用于总学分
+
 /**
- * 学分统计视图实体类 (CreditSummary)
- * 对应数据库中的 CreditSummary 视图，用于展示学生修读的总学分。
+ * CreditSummary 模型类 (数据传输对象 DTO)
+ * 代表从 CreditSummary 视图中查询到的学生学分统计信息。
+ * 这个类不是直接对应一个可修改的表，而是用于展示查询结果。
  */
 public class CreditSummary {
 
-    private int studentId; // 学生ID
-    private String studentName; // 学生姓名 (对应视图中的 s.name AS student_name)
-    private double totalCredits; // 学生获得的总学分 (对应视图中的 SUM(c.credit) AS total_credits)
+    private Integer studentId;        // 学生ID
+    private String studentName;       // 学生姓名
+    private BigDecimal totalCredits;  // 该学生已获得的总学分
 
     // 默认构造函数
     public CreditSummary() {
     }
 
-    // 包含所有字段的构造函数 (可选)
-    public CreditSummary(int studentId, String studentName, double totalCredits) {
+    // 包含所有字段的构造函数
+    public CreditSummary(Integer studentId, String studentName, BigDecimal totalCredits) {
         this.studentId = studentId;
         this.studentName = studentName;
         this.totalCredits = totalCredits;
     }
 
     // Getter 和 Setter 方法
-    public int getStudentId() {
+    public Integer getStudentId() {
         return studentId;
     }
 
-    public void setStudentId(int studentId) {
+    public void setStudentId(Integer studentId) {
         this.studentId = studentId;
     }
 
-    public String getStudentName() { // Getter 方法名更新为 getStudentName
+    public String getStudentName() {
         return studentName;
     }
 
-    public void setStudentName(String studentName) { // Setter 方法名更新为 setStudentName，参数也更新
+    public void setStudentName(String studentName) {
         this.studentName = studentName;
     }
 
-    public double getTotalCredits() {
+    public BigDecimal getTotalCredits() {
         return totalCredits;
     }
 
-    public void setTotalCredits(double totalCredits) {
+    public void setTotalCredits(BigDecimal totalCredits) {
         this.totalCredits = totalCredits;
     }
